backend/requirements.txt
langchain==0.3.25
langchain-groq==0.2.4
langgraph==0.4.1
python-dotenv==1.1.0
